requirements.txt
aiogram==3.13.1
request==2.32.3